pyproject.toml
[tool.poetry]
name = "whatsapp-job-agent"
version = "0.1.0"
description = "WhatsApp Job Agent v2"
authors = ["Pranay <pranay@example.com>"]
readme = "README.md"
packages = [
    { include = "job_backend", from = "apps/backend/src" },
    { include = "job_agent_runtime", from = "apps/agent-runtime/src" },
    { include = "job_platform", from = "packages/platform/src" },
    { include = "job_integrations", from = "packages/integrations/src" },
]

[tool.poetry.dependencies]
python = "^3.11"
anthropic = ">=0.84.0"
asyncpg = ">=0.29.0"
sqlalchemy = { version = ">=2.0.30", extras = ["asyncio"] }
alembic = ">=1.13.0"
httpx = ">=0.27.0"
google-auth = ">=2.28.0"
google-auth-oauthlib = ">=1.2.0"
google-auth-httplib2 = ">=0.2.0"
google-api-python-client = ">=2.120.0"
python-docx = ">=1.1.0"
pydantic = ">=2.7.0"
pydantic-settings = ">=2.3.0"
python-dotenv = ">=1.0.0"
schedule = ">=1.2.0"
structlog = ">=24.0.0"
rich = ">=13.7.0"
opentelemetry-sdk = ">=1.24.0"
opentelemetry-exporter-otlp = ">=1.24.0"
fastapi = ">=0.111.0"
uvicorn = ">=0.30.0"
psycopg2-binary = ">=2.9.9"
pypdf = ">=5.2.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
ruff = ">=0.4.0"
black = ">=24.0.0"
mypy = ">=1.10.0"
respx = ">=0.21.0"
factory-boy = ">=3.3.0"

[tool.pytest.ini_options]
markers = [
    "live_waha: live WAHA integration tests (opt-in with RUN_LIVE_WAHA_TESTS=1)",
    "live_anthropic: live Anthropic manager relevance tests (opt-in with RUN_LIVE_ANTHROPIC_TESTS=1)",
]

[tool.poetry.scripts]
job-backend = "job_backend.main:main"
job-agent-runtime = "job_agent_runtime.main:main"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
